requirements.txt
# ── API Clients ────────────────────────────────────
groq>=0.9.0
google-genai>=0.7.0
openai>=1.30.0
python-dotenv>=1.0.0

# ── ML / NLP (Local Models) ───────────────────────
torch>=2.2.0
transformers>=4.40.0
sentence-transformers>=3.0.0
openai-whisper>=20240930

# ── Vector Database ───────────────────────────────
chromadb>=0.5.0

# ── Market Data ───────────────────────────────────
yfinance>=0.2.40
nsepython>=2.9

# ── Technical Analysis ────────────────────────────
pandas-ta>=0.3.14b
backtesting>=0.3.3

# ── Chart Rendering ──────────────────────────────
mplfinance>=0.12.10b0
matplotlib>=3.8.0
Pillow>=10.0.0

# ── Backend ──────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0

# ── Data Processing ─────────────────────────────
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
pytz>=2024.1
beautifulsoup4>=4.12.0
lxml>=5.2.0

# ── Utilities ────────────────────────────────────
tqdm>=4.66.0
python-multipart>=0.0.9
